network/bootp: PR2031: Add and use header file
