+++ cpukit/libnetworking/rtems/rtems_bootp.c
@@ -12,6 +12,8 @@
 #include <rtems/rtems_bsdnet.h>
 #include <rtems/rtems_bsdnet_internal.h>
 
+#include "rtems/bootp.h"
+
 /*
  * Perform a BOOTP request
  */
