+++ cpukit/libnetworking/rtems/rtems_bsdnet_internal.h
@@ -179,7 +179,6 @@ void domaininit (void *);
 void ifinit (void *);
 void ipintr (void);
 void arpintr (void);
-bool bootpc_init(bool, bool);
 int socket (int, int, int);
 int ioctl (int, ioctl_command_t, ...);
